Update README.md
This project is a simple yet powerful Personal Budget Tracker built entirely in Microsoft Excel. It allows users to manage and track their monthly income and expenses using structured data tables, automated calculations, visual charts, and pivot table analysis.

The goal of this project is to demonstrate how Excel can be used not just for data entry, but also for financial analysis, data visualization, and decision making — all using tools that are accessible to almost everyone.